Added KeyRemovalManager Class to manage the LinkedBlockingDeque, Updated TTL Object, Renamed HashMap

+++ src/cache_lib/CacheLibrary.java
@@ -4,7 +4,7 @@ import java.util.Set;
 
 public interface CacheLibrary<K, V> {
     void put(K key, V value);
-    void put(K key, V value, int ttlDuration);
+    void put(K key, V value, long ttlDuration);
     V get(K key);
     boolean remove(K key);
     int size();
